requirements.txt
aioredis==2.0.0
anyio==4.3.0
fastapi==0.110.0
fastapi-cache2==0.2.1
numpy==1.24.4
openai==1.16.2
pandas==2.0.3
protobuf==5.26.0
pydantic==2.6.4
PyMySQL==1.1.0
uvicorn==0.29.0
cryptography==42.0.5
PyPDF2==3.0.1
redis==5.0.3
python-docx==1.1.0
requests==2.31.0
uvicorn==0.29.0
PyQRCode==1.2.1
qrcode==8.0
pypng
